sync on 2026/08/16, rev f4628801f5a283709811c05d07722dcf7cf408ba

+++ prog/1stPartyLibs/daScript/include/das_hash_map/das_hash_map.h
@@ -538,9 +538,9 @@ private:
         for ( size_t i = 0, n = hashes_.size(); i < n; ++i ) {
             const uint64_t kh = hashes_[i];
             if ( kh <= daslang_hash_map_detail::HASH_KILLED ) continue;
-            uint64_t j = kh & mask;
+            size_t j = size_t(kh & mask);
             while ( new_hashes[j] != daslang_hash_map_detail::HASH_EMPTY ) {
-                j = daslang_hash_map_detail::next_probe_slot(j, mask);
+                j = size_t(daslang_hash_map_detail::next_probe_slot(j, mask));
             }
             new_hashes[j] = kh;
             new_keys  [j] = das::move(keys_  [i]);
@@ -851,9 +851,9 @@ private:
         for ( size_t i = 0, n = hashes_.size(); i < n; ++i ) {
             const uint64_t kh = hashes_[i];
             if ( kh == daslang_hash_map_detail::HASH_EMPTY ) continue;
-            uint64_t j = kh & mask;
+            size_t j = size_t(kh & mask);
             while ( new_hashes[j] != daslang_hash_map_detail::HASH_EMPTY ) {
-                j = daslang_hash_map_detail::next_probe_slot(j, mask);
+                j = size_t(daslang_hash_map_detail::next_probe_slot(j, mask));
             }
             new_hashes[j] = kh;
             new_keys  [j] = das::move(keys_  [i]);
@@ -1074,9 +1074,9 @@ private:
         for ( size_t i = 0, n = hashes_.size(); i < n; ++i ) {
             const uint64_t kh = hashes_[i];
             if ( kh <= daslang_hash_map_detail::HASH_KILLED ) continue;
-            uint64_t j = kh & mask;
+            size_t j = size_t(kh & mask);
             while ( new_hashes[j] != daslang_hash_map_detail::HASH_EMPTY ) {
-                j = daslang_hash_map_detail::next_probe_slot(j, mask);
+                j = size_t(daslang_hash_map_detail::next_probe_slot(j, mask));
             }
             new_hashes[j] = kh;
             new_keys  [j] = das::move(keys_[i]);
@@ -1265,9 +1265,9 @@ private:
         for ( size_t i = 0, n = hashes_.size(); i < n; ++i ) {
             const uint64_t kh = hashes_[i];
             if ( kh == daslang_hash_map_detail::HASH_EMPTY ) continue;
-            uint64_t j = kh & mask;
+            size_t j = size_t(kh & mask);
             while ( new_hashes[j] != daslang_hash_map_detail::HASH_EMPTY ) {
-                j = daslang_hash_map_detail::next_probe_slot(j, mask);
+                j = size_t(daslang_hash_map_detail::next_probe_slot(j, mask));
             }
             new_hashes[j] = kh;
             new_keys  [j] = das::move(keys_[i]);
